+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Use new minisign identity proof generation algorithm.
+
 ## [1.30.0] - 2023-07-18
 
 ### Fixed