+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-### Changed
+### Added
 
 - Use new minisign identity proof generation algorithm.
+- Redesign minisign identity proof generation page.
 
 ## [1.30.0] - 2023-07-18