+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Use new minisign identity proof generation algorithm.
 - Redesign minisign identity proof generation page.
+- Added gallery page.
 
 ### Changed