+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.31.0] - 2023-07-26
+
 ### Added
 
 - Use new minisign identity proof generation algorithm.